requirements.txt
absl-py==2.2.2
accelerate==1.5.2
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
av==12.0.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clip==0.2.0
comm==0.2.2
contourpy==1.3.1
crc32c==2.7.1
cupy-cuda12x==13.6.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.15.4
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
et-xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.44
gradio==5.22.0
gradio-client==1.8.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.10
hjson==3.1.0
# hope==3.6.5
httpcore==1.0.7
httpx==0.28.1
huggingface==0.0.1
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
inflect==7.5.0
inquirerpy==0.3.4
ipdb==0.13.13
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.14
kiwisolver==1.4.8
levenshtein==0.27.1
loguru==0.7.2
lpips==0.1.4
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook==7.4.0
notebook-shim==0.2.4
numpy==2.2.5
omegaconf==2.3.0
openai==1.103.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.1.0
piq==0.8.0
platformdirs==4.3.7
portalocker==3.1.1
prettytable==3.16.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.1
psutil==7.0.0
ptwt==0.1.9
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.9
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytz==2025.2
pywavelets==1.8.0
pyyaml==6.0.2
pyzmq==26.4.0
rapidfuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screen==1.0.1
seaborn==0.13.2
semantic-version==2.10.0
send2trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
sty==1.0.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tfrecord==1.14.5
threadpoolctl==3.6.0
thrift==0.21.0
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.15.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
triton==3.1.0
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.0
validators==0.34.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
widgetsnbextension==4.0.14
word2number==1.1
xlsxwriter==3.2.2
yarl==1.19.0
zipp==3.21.0
datasets==3.5.0
TorchCodec==0.1.0
